+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 6.0.0 (unreleased)
 
+- Changed SQL to return dates instead of times for day, week, month, quarter, and year
 - Added `n` option for Redshift
 - Dropped support for Ruby < 2.6 and Rails < 5.2